+++ client/src/components/categories/DropdownMenu.jsx
@@ -1,0 +1,43 @@
+import React from "react";
+import "./DropdoenMenu.css";
+const DropdownMenu = () => {
+  return (
+    <>
+      <div class="topnav" id="myTopnav">
+        <a href="#home" class="active">
+          Home
+        </a>
+        <a href="#news">News</a>
+        <a href="#contact">Contact</a>
+        <div class="dropdown">
+          <button class="dropbtn">
+            Dropdown
+            <i class="fa fa-caret-down"></i>
+          </button>
+          <div class="dropdown-content">
+            <a href="#">Link 1</a>
+            <a href="#">Link 2</a>
+            <a href="#">Link 3</a>
+          </div>
+        </div>
+        <a href="#about">About</a>
+        <a
+          href="javascript:void(0);"
+          style="font-size:15px;"
+          class="icon"
+          onclick="myFunction()"
+        >
+          &#9776;
+        </a>
+      </div>
+
+      <div style="padding-left:16px">
+        <h2>Responsive Topnav with Dropdown</h2>
+        <p>Resize the browser window to see how it works.</p>
+        <p>Hover over the dropdown button to open the dropdown menu.</p>
+      </div>
+    </>
+  );
+};
+
+export default DropdownMenu;
