+++ client/src/components/header/Header.css
@@ -27,6 +27,17 @@
     margin-bottom: -2rem!important;
     margin-left: 1rem!important;
 }
+
+.modifymenu{
+
+    position: absolute;
+    transform: translate3d(-239px, 35px, 0px);
+    top: -0.5rem!important;
+    left: 0px;
+    will-change: transform;
+    width: 300px;
+    height: 51.5rem;
+}
 @media screen and (max-width: 780px) {
 
     .navbarwith{
