+++ client/src/components/slider/Silderuse.jsx
@@ -1,16 +1,42 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
+import axios from "axios";
 import "./Slider.css";
 const delay = 2500;
 
-function Slideruse({ image }) {
+function Slideruse() {
   const [index, setIndex] = React.useState(0);
   const timeoutRef = React.useRef(null);
+  const [image, setimage] = useState("");
 
   function resetTimeout() {
     if (timeoutRef.current) {
       clearTimeout(timeoutRef.current);
     }
   }
+  const getsilderimg = async () => {
+    axios.defaults.headers.post[
+      "Authorization"
+    ] = `Bearer ${localStorage.getItem("token")}`;
+    const config = {
+      headers: {
+        "Content-Type": "application/json",
+      },
+    };
+
+    const data = await axios.get(
+      `${process.env.REACT_APP_URL}/api/admin/getslider`,
+
+      config
+    );
+
+    console.log("img data from home", data.data.images);
+
+    setimage(data.data.images);
+  };
+
+  useEffect(() => {
+    getsilderimg();
+  }, []);
 
   React.useEffect(() => {
     resetTimeout();
