+++ client/src/components/slider/Slider.css
@@ -50,7 +50,7 @@
 
   
 @media screen and (max-width: 780px) {
-   
+  
   .slide {
     display: inline-block;
   
@@ -67,8 +67,5 @@
    
   }
 
-  .slideshow {
-    
-    margin-top: 12rem!important;
-  }
+  
   }
